requirements.txt
numpy==1.24.3
pydicom==2.3.1
pygame==2.4.0
PyOpenGL
PyOpenGL-accelerate
